ImageValZeroShot/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
qwen-vl-utils
Pillow>=8.0.0
accelerate>=0.20.0
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.62.0